Decouple presets from CollectionViewController and create Keyboard cell type
