+++ EyeSpeak/Features/Presets/PresetUICollectionViewCompositionalLayout.swift
@@ -10,6 +10,9 @@ import UIKit
 
 class PresetUICollectionViewCompositionalLayout: UICollectionViewCompositionalLayout {
     
+    private static let totalHeight: CGFloat = 834.0
+    private static let totalWidth: CGFloat = 1112.0
+    
     var dataSource: UICollectionViewDiffableDataSource<PresetsViewController.Section, PresetsViewController.ItemWrapper>? {
         self.collectionView?.dataSource as? UICollectionViewDiffableDataSource<PresetsViewController.Section, PresetsViewController.ItemWrapper>
     }
@@ -32,5 +35,144 @@ class PresetUICollectionViewCompositionalLayout: UICollectionViewCompositionalLa
         }
         attr?.transform = CGAffineTransform(translationX: 0, y: 500.0)
         return attr
+    }
+    
+    // MARK: - Section Layouts
+    
+    static func topBarSectionLayout() -> NSCollectionLayoutSection {
+        let redoButtonItem = NSCollectionLayoutItem(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(116.0 / 240.0),
+                                               heightDimension: .fractionalHeight(1.0)))
+        
+        let keyboardButtonItem = NSCollectionLayoutItem(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(116.0 / 240.0),
+                                               heightDimension: .fractionalHeight(1.0)))
+        
+        let subitems = [redoButtonItem, keyboardButtonItem]
+        
+        let containerGroup = NSCollectionLayoutGroup.horizontal(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(1),
+                                               heightDimension: .fractionalHeight(64.0 / totalHeight)),
+            subitems: subitems)
+        containerGroup.interItemSpacing = .fixed(8)
+        
+        let section = NSCollectionLayoutSection(group: containerGroup)
+        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 256, bottom: 0, trailing: 256)
+        
+        return section
+    }
+    
+    static func textFieldSectionLayout() -> NSCollectionLayoutSection {
+        let textFieldItem = NSCollectionLayoutItem(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0),
+                                               heightDimension: .fractionalHeight(1.0)))
+        
+        let subitems = [textFieldItem]
+        
+        let containerGroup = NSCollectionLayoutGroup.horizontal(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(1),
+                                               heightDimension: .fractionalHeight(86.0 / totalHeight)),
+            subitems: subitems)
+        containerGroup.interItemSpacing = .flexible(0)
+        containerGroup.edgeSpacing = NSCollectionLayoutEdgeSpacing(leading: .flexible(400), top: .fixed(0), trailing: .flexible(400), bottom: .fixed(0))
+        
+        let section = NSCollectionLayoutSection(group: containerGroup)
+        section.contentInsets = NSDirectionalEdgeInsets(top: 16, leading: 32, bottom: 0, trailing: 32)
+        
+        return section
+    }
+    
+    static func categoriesSectionLayout() -> NSCollectionLayoutSection {
+        let category1Item = NSCollectionLayoutItem(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(262.0 / 1048.0),
+                                               heightDimension: .fractionalHeight(1.0)))
+        let category2Item = NSCollectionLayoutItem(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(262.0 / 1048.0),
+                                               heightDimension: .fractionalHeight(1.0)))
+        let category3Item = NSCollectionLayoutItem(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(262.0 / 1048.0),
+                                               heightDimension: .fractionalHeight(1.0)))
+        let category4Item = NSCollectionLayoutItem(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(262.0 / 1048.0),
+                                               heightDimension: .fractionalHeight(1.0)))
+        
+        let subitems = [category1Item, category2Item, category3Item, category4Item]
+        
+        let containerGroup = NSCollectionLayoutGroup.horizontal(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(1),
+                                               heightDimension: .fractionalHeight(137.0 / totalHeight)),
+            subitems: subitems)
+        containerGroup.contentInsets = .init(top: 8, leading: 8, bottom: 8, trailing: 8)
+        let section = NSCollectionLayoutSection(group: containerGroup)
+        
+        let backgroundDecoration = NSCollectionLayoutDecorationItem.background(elementKind: "CategorySectionBackground")
+        backgroundDecoration.contentInsets = NSDirectionalEdgeInsets(top: 16, leading: 32, bottom: 16, trailing: 32)
+        
+        section.decorationItems = [backgroundDecoration]
+        section.contentInsets = NSDirectionalEdgeInsets(top: 16, leading: 32, bottom: 16, trailing: 32)
+        return section
+    }
+    
+    static func presetsSectionLayout() -> NSCollectionLayoutSection {
+        let presetItem = NSCollectionLayoutItem(layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(339.0 / totalWidth),
+                                                                                   heightDimension: .fractionalHeight(1.0)))
+        
+        let rowGroup = NSCollectionLayoutGroup.horizontal(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(1),
+                                               heightDimension: .fractionalHeight(1.0)),
+            subitem: presetItem, count: 3)
+        rowGroup.interItemSpacing = .fixed(16)
+        
+        let containerGroup = NSCollectionLayoutGroup.vertical(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(1),
+                                               heightDimension: .fractionalHeight(464.0 / totalHeight)),
+            subitem: rowGroup, count: 3)
+        containerGroup.interItemSpacing = .fixed(16)
+        containerGroup.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 32, bottom: 0, trailing: 32)
+        
+        let section = NSCollectionLayoutSection(group: containerGroup)
+        section.interGroupSpacing = 0
+        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 0)
+        section.orthogonalScrollingBehavior = .groupPaging
+        
+        let headerFooterSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0), heightDimension: .absolute(44))
+        let footer = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: headerFooterSize, elementKind: UICollectionView.elementKindSectionFooter, alignment: .bottom)
+        footer.extendsBoundary = true
+        footer.pinToVisibleBounds = true
+        section.boundarySupplementaryItems = [footer]
+        
+        return section
+    }
+    
+    static func keyboardSectionLayout() -> NSCollectionLayoutSection {
+        // TODO
+        let groupedKeyboardKeysItem = NSCollectionLayoutItem(layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(339.0 / totalWidth),
+                                                                                            heightDimension: .fractionalHeight(1.0)))
+        
+        let rowGroup = NSCollectionLayoutGroup.horizontal(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(1),
+                                               heightDimension: .fractionalHeight(1.0)),
+            subitem: groupedKeyboardKeysItem, count: 3)
+        rowGroup.interItemSpacing = .fixed(8)
+        
+        let containerGroup = NSCollectionLayoutGroup.vertical(
+            layoutSize: NSCollectionLayoutSize(widthDimension: .fractionalWidth(1),
+                                               heightDimension: .fractionalHeight(464.0 / totalHeight)),
+            subitem: rowGroup, count: 3)
+        containerGroup.interItemSpacing = .fixed(16)
+        containerGroup.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 32, bottom: 0, trailing: 32)
+        
+        let section = NSCollectionLayoutSection(group: containerGroup)
+        section.interGroupSpacing = 0
+        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 0)
+        section.orthogonalScrollingBehavior = .groupPaging
+        
+        let headerFooterSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0), heightDimension: .absolute(44))
+        let footer = NSCollectionLayoutBoundarySupplementaryItem(layoutSize: headerFooterSize, elementKind: UICollectionView.elementKindSectionFooter, alignment: .bottom)
+        footer.extendsBoundary = true
+        footer.pinToVisibleBounds = true
+        section.boundarySupplementaryItems = [footer]
+        
+        return section
     }
 }
